% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{abstract_class_error}
\alias{abstract_class_error}
\title{Abstract Class Error}
\usage{
abstract_class_error(cls_name, call = sys.call(-1), ...)
}
\arguments{
\item{cls_name}{name of class that was instantiated.}

\item{call}{function call to show in the condition.}

\item{...}{var args}
}
\value{
error condition
}
\description{
Abstract Class Error
}
\examples{

ABC <- R6::R6Class(
  classname = "ABC",
  public = list(
    initialize = function() {

      cls_name <- get_class_name()

      if(cls_name == "ABC")
        abstract_class_error(cls_name, call = "initialize")
    }
  )
)

result <- tryCatch(expr = {
  ABC$new()
},
error = function(e) {
 message(e)
 NULL
})
}
