% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{App}
\alias{App}
\title{Global Application}
\format{
An object of class \code{Application} (inherits from \code{R6}) of length 3.
}
\usage{
App
}
\description{
Global Application
}
\keyword{datasets}
