% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application.R
\docType{class}
\name{Application}
\alias{Application}
\title{Application

R6 Class representing the shiny application.}
\description{
The application instance.
}
\details{
This is a singleton (locked) instance that represents the
application. This variable is used to perform application
level (global) tasks, such as manage resources and
register pages for dispatch.
}
\concept{Application}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Application$new()}}
\item \href{#method-clone}{\code{Application$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of a log config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Application} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Application$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
