% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{as.character.LogLevel}
\alias{as.character.LogLevel}
\title{get_name}
\usage{
\method{as.character}{LogLevel}(x, ...)
}
\arguments{
\item{x}{log level}

\item{...}{ignored}
}
\value{
log level name
}
\description{
gets the name of the log level.
}
