% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{as.integer.LogLevel}
\alias{as.integer.LogLevel}
\title{Gets the severity of a log level.}
\usage{
\method{as.integer}{LogLevel}(x, ...)
}
\arguments{
\item{x}{log level}

\item{...}{ignored}
}
\value{
log level
}
\description{
Gets the severity of a log level.
}
