% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\docType{class}
\name{Control}
\alias{Control}
\title{Control}
\description{
An object that is meant to be the base type for
all ui controls. Controls are meant to represent
self-contained functional units (e.g., charts,
menus, tables, etc.).
}
\details{
Controls are modularized containers for UI widgets.
}
\section{Super class}{
\code{\link[shiny.gentelella:Element]{shiny.gentelella::Element}} -> \code{Control}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Control$new()}}
\item \href{#method-clone}{\code{Control$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="class_name">}\href{../../shiny.gentelella/html/Element.html#method-class_name}{\code{shiny.gentelella::Element$class_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="identifier">}\href{../../shiny.gentelella/html/Element.html#method-identifier}{\code{shiny.gentelella::Element$identifier()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this
\link{Element} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Control$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Control} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Control$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
