% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{Dashboard}
\alias{Dashboard}
\title{Dashboard}
\description{
An object that is meant to represent a standardized
dashboard page layout with \link{NavigationMenu}, Settings,
Summary and a customizable container for the main
content section.
}
\references{
\code{\link{Page}}
}
\section{Super classes}{
\code{\link[shiny.gentelella:Element]{shiny.gentelella::Element}} -> \code{\link[shiny.gentelella:Page]{shiny.gentelella::Page}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dashboard$new()}}
\item \href{#method-layout}{\code{Dashboard$layout()}}
\item \href{#method-clone}{\code{Dashboard$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="class_name">}\href{../../shiny.gentelella/html/Element.html#method-class_name}{\code{shiny.gentelella::Element$class_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="identifier">}\href{../../shiny.gentelella/html/Element.html#method-identifier}{\code{shiny.gentelella::Element$identifier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Page" data-id="page">}\href{../../shiny.gentelella/html/Page.html#method-page}{\code{shiny.gentelella::Page$page()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Page" data-id="server">}\href{../../shiny.gentelella/html/Page.html#method-server}{\code{shiny.gentelella::Page$server()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a portal page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-layout"></a>}}
\if{latex}{\out{\hypertarget{method-layout}{}}}
\subsection{Method \code{layout()}}{
Wrapper around a dashboard templated
layout.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$layout()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dashboard$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
