% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element.R
\docType{class}
\name{Element}
\alias{Element}
\title{Element

R6 Class representing a UI Element}
\value{
top-level class name

element's uuid
}
\description{
This is the abstract base class for all UI
elements.
}
\details{
An object that is meant to be the base type for
all user-interface elements. The primary driver
for having a base type for all UI elements is
that every ui should have a unique id for
shiny module dispatch, and basic descriptors
that assist with robust logging.
}
\concept{Controls}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Element$new()}}
\item \href{#method-class_name}{\code{Element$class_name()}}
\item \href{#method-identifier}{\code{Element$identifier()}}
\item \href{#method-clone}{\code{Element$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Element} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-class_name"></a>}}
\if{latex}{\out{\hypertarget{method-class_name}{}}}
\subsection{Method \code{class_name()}}{
Get accessor for (top-level) class name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$class_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identifier"></a>}}
\if{latex}{\out{\hypertarget{method-identifier}{}}}
\subsection{Method \code{identifier()}}{
Get accessor for object's uuid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$identifier()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Element$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
