% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{format.log_level}
\alias{format.log_level}
\title{Get log level message format}
\usage{
\method{format}{log_level}(level, msg)
}
\arguments{
\item{level}{log level}

\item{msg}{log msg}
}
\value{
formatted log message
}
\description{
Get log level message format
}
