% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{get_class_name}
\alias{get_class_name}
\title{get_class_name}
\usage{
get_class_name()
}
\value{
run-time class name
}
\description{
Used to determine the run-time class name of
an object.
}
\details{
This function determines the type of the instantiated
object at run-time, which has a tremdious amount of
utility, especially objets with deeply nested inheritance.
Must be called within the object's constructor (\link{initialize}).
}
\examples{
Base <- R6::R6Class(
  classname = "Base",

  public = list(
    initialize = function() {
      private$.cls_name <- get_class_name()
    },

   class_name = function() {
      private$.cls_name
    }
  ),

  private = list(
    .cls_name = NULL
  )
)
Derived <- R6::R6Class(
  classname = "Derived",
  inherit = Base,

  public = list(),
  private = list()
)

base <- Base$new(); derived <- Derived$new()

print(base$class_name())
print(derived$class_name())
}
