% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{get_level.LogLevel}
\alias{get_level.LogLevel}
\title{get_name}
\usage{
get_level.LogLevel(level)
}
\arguments{
\item{level}{log level}
}
\value{
log level name
}
\description{
gets the name of the log level.
}
