% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{get_level_name.log_level}
\alias{get_level_name.log_level}
\title{Get Level Name}
\usage{
\method{get_level_name}{log_level}(level)
}
\arguments{
\item{level}{log level}
}
\value{
log level name
}
\description{
gets the name of the log level.
}
