% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{get_level_name}
\alias{get_level_name}
\title{Get Level Name}
\usage{
get_level_name(level)
}
\arguments{
\item{level}{log level}
}
\value{
log level name
}
\description{
gets the name of the log level.
}
