% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{get_level_severity}
\alias{get_level_severity}
\title{get level severity}
\usage{
get_level_severity(level)

get_level_severity(level)
}
\arguments{
\item{level}{log level}
}
\value{
level severity


}
\description{
get level severity

Gets the severity of a log level.
}
\examples{
get_severity(FATAL)
}
