% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{get_severity}
\alias{get_severity}
\title{get_level}
\usage{
get_severity(level)

get_severity(level)
}
\arguments{
\item{level}{log level}
}
\value{
log level


}
\description{
get_level

Gets the severity of a log level.
}
\examples{
get_severity(FATAL)
}
