% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{level_info}
\alias{level_info}
\title{Gets style level information}
\usage{
level_info(level)

level_info(level)

level_info(level)
}
\arguments{
\item{level}{log level}
}
\value{
styled level information

styled level information

styled level information
}
\description{
Gets style level information

Gets style level information

Gets style level information
}
