% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{level_severity}
\alias{level_severity}
\title{Gets the severity of a log level.}
\usage{
level_severity(level)
}
\arguments{
\item{level}{log level}
}
\value{

}
\description{
Gets the severity of a log level.
}
\examples{
get_severity(FATAL)
}
