% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-config.R
\name{LogConfig}
\alias{LogConfig}
\title{R6 Class to encapsulate common logging configurations.}
\value{
customized log layout for class instance.
}
\description{
R6 Class to encapsulate common logging configurations.

R6 Class to encapsulate common logging configurations.
}
\details{
An object to encapsulate common logging  configurations,
formatters, generators, colorizers, etc. Designed to
be used via composition inside UI elements (pages,
controls, etc) as parent object is expected to have
UI Element interface (id, cls_name).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogConfig$new()}}
\item \href{#method-generate_layout}{\code{LogConfig$generate_layout()}}
\item \href{#method-clone}{\code{LogConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of a log config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogConfig$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{LogConfig} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_layout"></a>}}
\if{latex}{\out{\hypertarget{method-generate_layout}{}}}
\subsection{Method \code{generate_layout()}}{
generates a customized logging
layout for passed in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogConfig$generate_layout(cls)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cls}}{class to generate log layout for.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
