% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-dispatch.R
\docType{class}
\name{LogDispatch}
\alias{LogDispatch}
\title{LogDispatch}
\description{
R6 Class that dispatches log messages throughout the application.
}
\details{
This object is designed to a centralized logging dispatcher that
renders log messages with the appropriate context of the calling
object. The \link{LogLayout} object is used to generate log message
layouts (render formats), which are used by the \link{LogDispatcher}
to render highly-customizable and detailed log messages.
}
\section{Metrics}{


System Context

\itemize{
\item{"sysname"} : {The operating system name.}
\item{"release"} : {The OS release.}
\item{"version"} : {The OS version.}
\item{"nodename"} : {A name by which the machine is known on the network (if any).}
\item{"machine"} : {A concise description of the hardware, often the CPU type.}
\item{"login"} : {The user's login name, or "unknown" if it cannot be ascertained.}
\item{"user"} : {The name of the real user ID, or "unknown" if it cannot be ascertained.}
\item{"r-ver"} : {R Version \link{major}.\link{minor}}
}
}

\seealso{
LogLevel

Other Logging: 
\code{\link{LogLayoutGenerator}},
\code{\link{LogLayout}}
}
\concept{Logging}
\section{Super class}{
\code{\link[shiny.gentelella:Singleton]{shiny.gentelella::Singleton}} -> \code{LogDispatch}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{system}}{returns a lazy-loaded field to store
system level logging metrics.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogDispatch$new()}}
\item \href{#method-trace}{\code{LogDispatch$trace()}}
\item \href{#method-info}{\code{LogDispatch$info()}}
\item \href{#method-debug}{\code{LogDispatch$debug()}}
\item \href{#method-warn}{\code{LogDispatch$warn()}}
\item \href{#method-error}{\code{LogDispatch$error()}}
\item \href{#method-fatal}{\code{LogDispatch$fatal()}}
\item \href{#method-succes}{\code{LogDispatch$succes()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Singleton" data-id="class_name">}\href{../../shiny.gentelella/html/Singleton.html#method-class_name}{\code{shiny.gentelella::Singleton$class_name()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of a log config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{LogLayout} object.
Log a trace level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trace"></a>}}
\if{latex}{\out{\hypertarget{method-trace}{}}}
\subsection{Method \code{trace()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$trace(msg, format)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}

\item{\code{format}}{log format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log an information level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$info(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log a debug level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug"></a>}}
\if{latex}{\out{\hypertarget{method-debug}{}}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$debug(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log a warn level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warn"></a>}}
\if{latex}{\out{\hypertarget{method-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$warn(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log an error level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-error"></a>}}
\if{latex}{\out{\hypertarget{method-error}{}}}
\subsection{Method \code{error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$error(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log a fatal level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fatal"></a>}}
\if{latex}{\out{\hypertarget{method-fatal}{}}}
\subsection{Method \code{fatal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$fatal(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
Log a success level message
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-succes"></a>}}
\if{latex}{\out{\hypertarget{method-succes}{}}}
\subsection{Method \code{succes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogDispatch$succes(msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{message to log}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
reference to self
}
}
}
