% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\docType{class}
\name{LogLayout}
\alias{LogLayout}
\title{LogLayout}
\value{
customized log layout for class instance.

log metric layout colorized by severity.

log metric layout colorized by severity.

log metric layout colorized by severity.
}
\description{
R6 Class to encapsulate common logging configurations.
}
\details{
An object to encapsulate common logging  configurations,
formatters, generators, colorizers, etc. Designed to
be used via composition inside UI elements (pages,
controls, etc) as parent object is expected to have
UI Element interface (id, cls_name).
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{LogLayoutGenerator}}
}
\concept{Logging}
\section{Super class}{
\code{\link[shiny.gentelella:Singleton]{shiny.gentelella::Singleton}} -> \code{LogLayout}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogLayout$new()}}
\item \href{#method-generate_ui_layout}{\code{LogLayout$generate_ui_layout()}}
\item \href{#method-format_metric_colorized_severity}{\code{LogLayout$format_metric_colorized_severity()}}
\item \href{#method-colorize_by_level}{\code{LogLayout$colorize_by_level()}}
\item \href{#method-grayscale_by_level}{\code{LogLayout$grayscale_by_level()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Singleton" data-id="class_name">}\href{../../shiny.gentelella/html/Singleton.html#method-class_name}{\code{shiny.gentelella::Singleton$class_name()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of a log config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayout$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{LogLayout} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_ui_layout"></a>}}
\if{latex}{\out{\hypertarget{method-generate_ui_layout}{}}}
\subsection{Method \code{generate_ui_layout()}}{
generates a customized logging
layout for passed in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayout$generate_ui_layout(cls)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cls}}{class to generate log layout for.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_metric_colorized_severity"></a>}}
\if{latex}{\out{\hypertarget{method-format_metric_colorized_severity}{}}}
\subsection{Method \code{format_metric_colorized_severity()}}{
formats a log metric by colorizing it based
on the log severity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayout$format_metric_colorized_severity(level, metric, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{log severity.}

\item{\code{metric}}{metric to log.}

\item{\code{...}}{additional styles}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-colorize_by_level"></a>}}
\if{latex}{\out{\hypertarget{method-colorize_by_level}{}}}
\subsection{Method \code{colorize_by_level()}}{
formats a log metric by colorizing it based
on the log severity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayout$colorize_by_level(level, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{log severity.}

\item{\code{metric}}{metric to log.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grayscale_by_level"></a>}}
\if{latex}{\out{\hypertarget{method-grayscale_by_level}{}}}
\subsection{Method \code{grayscale_by_level()}}{
formats a log metric by colorizing it based
on the log severity in gray scale format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayout$grayscale_by_level(level, metric)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{log severity.}

\item{\code{metric}}{metric to log.}
}
\if{html}{\out{</div>}}
}
}
}
