% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout-generator.R
\docType{class}
\name{LogLayoutGenerator}
\alias{LogLayoutGenerator}
\title{LogLayoutGenerator}
\description{
R6 Class to encapsulate common logging configurations.
}
\details{
An object to encapsulate common logging  configurations,
formatters, generators, colorizers, etc. Designed to
be used via composition inside UI elements (pages,
controls, etc) as parent object is expected to have
UI Element interface (id, cls_name).
}
\seealso{
Other Logging: 
\code{\link{LogDispatch}},
\code{\link{LogLayout}}
}
\concept{Logging}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogLayoutGenerator$new()}}
\item \href{#method-add_metric}{\code{LogLayoutGenerator$add_metric()}}
\item \href{#method-get_layout}{\code{LogLayoutGenerator$get_layout()}}
\item \href{#method-clone}{\code{LogLayoutGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of a log config.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayoutGenerator$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{LogLayoutGenerator} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_metric"></a>}}
\if{latex}{\out{\hypertarget{method-add_metric}{}}}
\subsection{Method \code{add_metric()}}{
appends a metric to the log layout.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayoutGenerator$add_metric(..., metric, env = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{layout style}

\item{\code{metric}}{log metric}

\item{\code{env}}{parent env}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A reference to self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_layout"></a>}}
\if{latex}{\out{\hypertarget{method-get_layout}{}}}
\subsection{Method \code{get_layout()}}{
returns the generated layout.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayoutGenerator$get_layout()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
generated layout.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLayoutGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
