% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation-menu.R
\name{NavigationMenu}
\alias{NavigationMenu}
\title{NavigationMenu}
\description{
R6 object that represents a navigation menu conrol.
}
\details{
This object is designed to generate a dashboard
page navigation menu. The file/resource that
is passed to the constructor should be a yml
file that specifies the structure of the menu.
}
\section{Super classes}{
\code{\link[shiny.gentelella:Element]{shiny.gentelella::Element}} -> \code{\link[shiny.gentelella:Control]{shiny.gentelella::Control}} -> \code{NavigationMenu}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NavigationMenu$new()}}
\item \href{#method-ui}{\code{NavigationMenu$ui()}}
\item \href{#method-clone}{\code{NavigationMenu$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="class_name">}\href{../../shiny.gentelella/html/Element.html#method-class_name}{\code{shiny.gentelella::Element$class_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="identifier">}\href{../../shiny.gentelella/html/Element.html#method-identifier}{\code{shiny.gentelella::Element$identifier()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a navigation menu based
on the structure of the yml
definition file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NavigationMenu$new(nav_definition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nav_definition}}{menu navigation file (yml)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{NavigationMenu} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
Constructs the navigation menu UI
from yml definition.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NavigationMenu$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NavigationMenu$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
