% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_fmt_line_break}
\alias{new_fmt_line_break}
\title{Formatted Line Break}
\usage{
new_fmt_line_break()
}
\value{
log layout newline.
}
\description{
Inserts a new line in the format.
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_log_layout}()},
\code{\link{value.log_layout}()}
}
\concept{Log Layout}
