% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_fmt_literal}
\alias{new_fmt_literal}
\title{Formatted Literal}
\usage{
new_fmt_literal(style, literal)
}
\arguments{
\item{literal}{log value}

\item{...}{format styles}
}
\value{
log metric layout.
}
\description{
Formatted Literal
}
\examples{
new_fmt_literal(red $ bold, "literal text")

new_fmt_literal(blue $ italic, "literal text")
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_metric}()},
\code{\link{new_line}()}
}
\concept{Log Layout}
