% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_line}
\alias{new_line}
\title{Formatted Line Break}
\usage{
new_line()
}
\value{
log layout newline.
}
\description{
Inserts a new line in the format.
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_metric}()}
}
\concept{Log Layout}
