% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_log_layout}
\alias{new_log_layout}
\title{Log Layout}
\usage{
new_log_layout(..., association = character())
}
\arguments{
\item{...}{collection of format objects to initialize with.}

\item{association}{objects to associate this log format with.}
}
\value{
object's value
}
\description{
a class that stores a collection of log format objects
and understands how to associate a given format to
a class of objects.
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_log_level}()},
\code{\link{new_fmt_metric}()},
\code{\link{value.log_layout}()}
}
\concept{Log Layout}
