% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-levels.R
\name{new_log_level}
\alias{new_log_level}
\title{LogLevel}
\usage{
new_log_level(name, severity)
}
\arguments{
\item{name}{name of the log level is the string representation.}

\item{severity}{log severity is used in threshold evaluations when
viewing a message of this type.}
}
\value{
new \link{log_level}
}
\description{
S3 object to represent a typed & predefined log level.
}
\section{Severity}{


\itemize{
\item{"FATAL"} : {The FATAL level designates very severe error events that will presumably lead the application to abort.}
\item{"ERROR"} : {The ERROR level designates error events that might still allow the application to continue running.}
\item{"WARN"} : {The WARN level designates potentially harmful situations.}
\item{"SUCCESS"} : {The SUCCESS level designates that the operation was unencumbered.}
\item{"INFO"} : {The INFO level designates informational messages that highlight the progress of the application at coarse-grained level.}
\item{"DEBUG"} : {The DEBUG Level designates fine-grained informational events that are most useful to debug an application.}
\item{"TRACE"} : {The TRACE Level designates finer-grained informational events than the DEBUG}
}
}

