% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{Page}
\alias{Page}
\title{Element}
\description{
An object that is meant to be the base type for
all ui pages. Pages are meant to represent an
entire page in the UI as serves as a container
for controls.
}
\references{
\code{\link{Element}}
}
\section{Super class}{
\code{\link[shiny.gentelella:Element]{shiny.gentelella::Element}} -> \code{Page}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Page$new()}}
\item \href{#method-page}{\code{Page$page()}}
\item \href{#method-server}{\code{Page$server()}}
\item \href{#method-clone}{\code{Page$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="class_name">}\href{../../shiny.gentelella/html/Element.html#method-class_name}{\code{shiny.gentelella::Element$class_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="identifier">}\href{../../shiny.gentelella/html/Element.html#method-identifier}{\code{shiny.gentelella::Element$identifier()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new UI Page object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Page} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-page"></a>}}
\if{latex}{\out{\hypertarget{method-page}{}}}
\subsection{Method \code{page()}}{
Wrapper around fluid layout page.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$page()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
server-side wrapper.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{form input}

\item{\code{output}}{render output}

\item{\code{session}}{page session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
