% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{resource_not_found_error}
\alias{resource_not_found_error}
\title{Resource Not Found Error}
\usage{
resource_not_found_error(resource_path, call = sys.call(-1), ...)
}
\arguments{
\item{resource_path}{path of the resource that was attempted to load.}

\item{call}{function call to show in the condition.}

\item{...}{var args}
}
\value{
error condition
}
\description{
Resource Not Found Error
}
