% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource-mgr.R
\name{ResourceMgr}
\alias{ResourceMgr}
\title{Resource Manager}
\value{
fully qualified file system path
}
\description{
R6 class that encapsulates common resource management
operations.
}
\details{
R6 class that is designed to be a singleton (there
is only a single instance globally) to provide
a clean wrapper over common resource management
tasks.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResourceMgr$new()}}
\item \href{#method-map_path}{\code{ResourceMgr$map_path()}}
\item \href{#method-map_resource}{\code{ResourceMgr$map_resource()}}
\item \href{#method-clone}{\code{ResourceMgr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new UI Page object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceMgr$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ResourceMgr} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_path"></a>}}
\if{latex}{\out{\hypertarget{method-map_path}{}}}
\subsection{Method \code{map_path()}}{
Maps a package path to a shiny resource path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceMgr$map_path(prefix, directory, warn_empty = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prefix}}{exposed path in the website}

\item{\code{directory}}{file system directory (mapped in a pkg)}

\item{\code{warn_empty}}{warn if no files are found.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_resource"></a>}}
\if{latex}{\out{\hypertarget{method-map_resource}{}}}
\subsection{Method \code{map_resource()}}{
Maps a package resource to a local file system resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceMgr$map_resource(path, package = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{file system patch}

\item{\code{package}}{package where the resource lives, if none
specified will use self.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceMgr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
