% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleton.R
\docType{class}
\name{Singleton}
\alias{Singleton}
\title{Singleton}
\description{
R6 class that encapsulates the process of setting up
an object as singleton instance.
}
\details{
A singleton is a common object-oriented design
pattern \link{https://en.wikipedia.org/wiki/Singleton_pattern}.
Even though some refer to this design as an 'anti-pattern, it
provides a great deal of utility, especially in the context
of building user-interface applications.
}
\examples{
```{r}
ExampleSingleton <- R6::R6Class(
classname = "ExampleSingleton",
inherit = Singleton,
public = list(

  initialize = function() {
    super$initialize()
  },

  append = function(item) {
    private$items = c(private$items, item)
  },

  clear = function() {
    private$items = list()
  },

  display = function() {
    cat(stringr::str_flatten(private$items, collapse = "\n"))
  }
),
active = list(),
private = list(
  items = list()
)
)

s1 <- ExampleSingleton$new(); s2 <- ExampleSingleton$new(); s3 <- ExampleSingleton$new()


s1$append("testing from inst 1")

s2$append("from inst 2")

s3$display()

s1$append("inst 1 again")

s2$display()

```
}
\concept{Design Patterns}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Singleton$new()}}
\item \href{#method-class_name}{\code{Singleton$class_name()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initializes a single instance of the type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singleton$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{Singleton} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-class_name"></a>}}
\if{latex}{\out{\hypertarget{method-class_name}{}}}
\subsection{Method \code{class_name()}}{
Gets the run-time name of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Singleton$class_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The \code{Singleton} object.
}
}
}
