% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{style.fmt_layout}
\alias{style.fmt_layout}
\title{Gets the style of a format object.}
\usage{
\method{style}{fmt_layout}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Gets the style of a format object.
}
