% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{style}
\alias{style}
\title{Gets the style of a format object.}
\usage{
style(fmt)
}
\arguments{
\item{fmt}{object to extract value from.}
}
\value{
object's value
}
\description{
Gets the style of a format object.
}
