% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{Table}
\alias{Table}
\title{Table}
\description{
An object that is designed to encapsulate
most of the standard plumbing in setting up
a Datatable in the UI. The html template
}
\references{
\code{\link{Page}}
}
\section{Super classes}{
\code{\link[shiny.gentelella:Element]{shiny.gentelella::Element}} -> \code{\link[shiny.gentelella:Control]{shiny.gentelella::Control}} -> \code{PortalTable}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{table title}

\item{\code{export_btn}}{export btn}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Table$new()}}
\item \href{#method-subtitle}{\code{Table$subtitle()}}
\item \href{#method-ui}{\code{Table$ui()}}
\item \href{#method-ui_extra}{\code{Table$ui_extra()}}
\item \href{#method-ui_footer}{\code{Table$ui_footer()}}
\item \href{#method-clone}{\code{Table$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="class_name">}\href{../../shiny.gentelella/html/Element.html#method-class_name}{\code{shiny.gentelella::Element$class_name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="shiny.gentelella" data-topic="Element" data-id="identifier">}\href{../../shiny.gentelella/html/Element.html#method-identifier}{\code{shiny.gentelella::Element$identifier()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize an instance of a
portal table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$new(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{title of the table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subtitle"></a>}}
\if{latex}{\out{\hypertarget{method-subtitle}{}}}
\subsection{Method \code{subtitle()}}{
function to customize the table subtitle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$subtitle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
render the ui for the a portal styled data table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_extra"></a>}}
\if{latex}{\out{\hypertarget{method-ui_extra}{}}}
\subsection{Method \code{ui_extra()}}{
pure virtual method for customizing the table UI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$ui_extra()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_footer"></a>}}
\if{latex}{\out{\hypertarget{method-ui_footer}{}}}
\subsection{Method \code{ui_footer()}}{
pure virtual method for customizing the table UI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$ui_footer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
