% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{unknown_severity_warning}
\alias{unknown_severity_warning}
\title{Unknown Severity Warning}
\usage{
unknown_severity_warning(level, call = sys.call(-1), ...)
}
\arguments{
\item{level}{attempted severity level of the log dispatch.}

\item{call}{function call to show in the condition.}

\item{...}{var args}
}
\value{
error condition
}
\description{
Unknown Severity Warning
}
