% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{value.fmt_metric}
\alias{value.fmt_metric}
\title{Gets the value of a format object.}
\usage{
\method{value}{fmt_metric}(fmt, ...)
}
\arguments{
\item{fmt}{object to extract value from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
object's value
}
\description{
Gets the value of a format object.
}
