% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_fmt_layout}
\alias{new_fmt_layout}
\title{Format Layout}
\usage{
new_fmt_layout(style, class = character())
}
\arguments{
\item{style}{\link{crayon} that the layout will use in log generation.}

\item{class}{class name}
}
\value{
new log format
}
\description{
Base type
}
\seealso{
Other Log Layout: 
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()},
\code{\link{new_fmt_metric}()}
}
\concept{Log Layout}
