% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-layout.R
\name{new_fmt_metric}
\alias{new_fmt_metric}
\title{Formatted Metric}
\usage{
new_fmt_metric(style, metric)
}
\arguments{
\item{style}{that the layout will use in log generation}

\item{metric}{the metric to log.}
}
\value{
a new formatted metric
}
\description{
Inserts a formatted log metric.
}
\examples{
new_fmt_metric(bold & red, "sysname")

new_fmt_metric(crayon::combine_styles(crayon::italic,
                                      crayon::make_style("darkorange")),
             "release")
}
\seealso{
\link{LogDispatch}

Other Log Layout: 
\code{\link{new_fmt_layout}()},
\code{\link{new_fmt_line_break}()},
\code{\link{new_fmt_literal}()}
}
\concept{Log Layout}
